+++ lib/titan.rb
@@ -1,6 +1,8 @@
 require "titan/version"
 
 module Titan
+  TITAN_DIRECTORY = File.expand_path('.titan_threads', '~')
+
   autoload :CLI,      "titan/cli"
   autoload :System,   "titan/system"
   autoload :Thread,   "titan/thread"