+++ spec/titan/system_spec.rb
@@ -14,4 +14,29 @@ describe Titan::System do
       end
     end
   end
+
+  describe ".check_file_system" do
+    before(:each) do
+      File.stub!(:directory?).and_return(false)
+    end
+
+    it "should create the directory" do
+      Dir.should_receive(:mkdir).with(Titan::TITAN_DIRECTORY)
+      Titan::System.check_filesystem
+    end
+  end
+
+  describe ".pid_files" do
+    it "should return an Array" do
+      Titan::System.pid_files.should be_an(Array)
+    end
+
+    it "should not include ." do
+      Titan::System.pid_files.should_not include(".")
+    end
+
+    it "should not include .." do
+      Titan::System.pid_files.should_not include("..")
+    end
+  end
 end